refactor(server): move service_* into services/ package
