+++ server/src/routes_search.py
@@ -1,6 +1,6 @@
 from flask import Blueprint, request, jsonify
 from config import logger, get_available_culling_presets
-import service_search
+from services import search as service_search
 
 search_bp = Blueprint("search", __name__)
