+++ server/src/routes_style_edit.py
@@ -17,9 +17,9 @@ from flask import Blueprint, jsonify, request
 
 from config import logger
 from routes_index import _extract_options, _extract_photo_ids
-import service_chroma as chroma_service
-import service_style_engine as style_engine
-from service_style_engine import CONFIDENCE_LOW
+from services import chroma as chroma_service
+from services import style_engine as style_engine
+from services.style_engine import CONFIDENCE_LOW
 from routes_edit import _persist_edit_recipe, _success_payload
 
 style_edit_bp = Blueprint("style_edit", __name__)
