+++ server/src/services/clip.py
