+++ server/src/services/exif.py
