+++ server/test/dump_photo_metadata.py
@@ -41,7 +41,7 @@ def _extract_photo_id(argv: list[str]) -> tuple[str | None, list[str]]:
 _PHOTO_ID, _NEW_ARGV = _extract_photo_id(sys.argv)
 sys.argv = _NEW_ARGV
 
-from service_chroma import get_all_image_ids, get_image  # noqa: E402  (import after argv tweak)
+from services.chroma import get_all_image_ids, get_image  # noqa: E402  (import after argv tweak)
 
 
 def main() -> int:
